doc/prog_guide: new chapter on cmdline library
The cmdline library was not documented in our programmers guide, so add
a new chapter on it. This chapter covers step-by-step how to use the
library, rather than focusing on the library internals. This complements
the existing cmdline example app document, providing more details on the
process of using the library.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>