net/ice: fix DCF port statistics not cleared
As the kernel driver does not reset the statistics after exiting testpmd,
the previous statistics being retained the next time testpmd is started.

Call 'ice_dcf_stats_reset' during the initialization of the DCF in
order to clear any statistics that may exist from the last use of the DCF
and to avoid statistics errors.

Fixes: 7564d5509611 ("net/ice: add DCF hardware initialization")
Cc: stable@dpdk.org

Signed-off-by: Zhichao Zeng <zhichaox.zeng@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>