net/ice: fix crash on closing representor ports
Since the representor port needs to access the resource of the
associated DCF when it is closing. Therefore, all the representor
port should be closed first, and then close the associated DCF port.

If the DCF port is closed before the representor port on PMD exit.
This will result in accessing freed resources and eventually a
core dump will occur.

This patch fixes this issue by notifying each other to unassociate
when the DCF port and the representor port are closed.

Fixes: 5674465a32c8 ("net/ice: add DCF VLAN handling")
Fixes: c7e1a1a3bfeb ("net/ice: refactor DCF VLAN handling")
Fixes: da9cdcd1f372 ("net/ice: fix crash on representor port closing")
Cc: stable@dpdk.org

Signed-off-by: Mingjin Ye <mingjinx.ye@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>