event/dlb2: fix disable PASID
In vfio-pci driver when PASID is enabled by default
DLB hardware puts DLB in SIOV mode. This breaks
DLB PF-PMD mode. For DLB PF-PMD mode to function properly
PASID needs to be disabled.

In this commit this issue is addressed and PASID is disabled
by writing a zero to PASID control register.

Fixes: 5433956d5185 ("event/dlb2: add eventdev probe")
Cc: stable@dpdk.org

Signed-off-by: Abdullah Sevincer <abdullah.sevincer@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>