ml/cnxk: fix fast-path synchronization
Segfaults are reported with TVM/LLVM models as NULL value
is read by dequeue thread for the op handle. This is due
synchronization issue between enqueue and dequeue threads.

This fix ensures the op handle written to internal request
structure is committed before the dequeue threads read the
handle value.

Fixes: 88001b4d88de ("ml/cnxk: update fast path functions")

Signed-off-by: Srikanth Yalavarthi <syalavarthi@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>