ethdev: add trace for the new fields added in 23.11
The 'max_rx_bufsize' and 'rss_algo_capa' in struct rte_eth_dev_info and
the 'algorithm' in struct rte_eth_rss_conf are introduced to 23.11.
This patch add the trace of these new fields in related trace point
interface.

In addition, the 'max_lro_pkt_size' in struct rte_eth_dev_info is also
important. So this patch also add it.

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Acked-by: Chengwen Feng <fengchengwen@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>