build: riscv is not a valid -march value
If building riscv natively with -Dplatform=generic config/meson.build
will select cpu_instruction_set=riscv.

That was fine because config/riscv/meson.build did override it to valid
values later, but since b7676fcccab4 ("config: verify machine arch
flag") it will break the build as it tries to test -march=riscv which
is not a value value.

The generic setting used in most cases is rv64gc, set this here
as well.

Fixes: b7676fcccab4 ("config: verify machine arch flag"
Fixes: f22e705ebf12 ("eal/riscv: support RISC-V architecture")

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>