arg_parser: new library for command line parsing
Add a new library to make it easier for eal and other libraries to parse
command line arguments.

The first function in this library is one to parse a corelist string into an
array of individual core ids. The function will then return the total number
of cores described in the corelist

Signed-off-by: Euan Bourke <euan.bourke@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>