common/qat: fix build errors with incompatible IPSec lib
Since "find_library()" is used to find the libipsec_mb library, there is
a chance that an incompatible library will be found e.g. when doing
cross-builds. To fix this, we copy the logic from the crypto/ipsec_mb
driver to actually do a link-check on the library before deciding it can
be used.

Fixes: ca0ba0e48129 ("crypto/qat: default to IPsec MB for computations")
Cc: brian.dooley@intel.com

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Acked-by: Ciara Power <ciara.power@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>