config/arm: use common cpu arch for cross files
The cpu info in some cross files is inconsistent with
that in SoC flags. The mismatch doesn't cause any issue
because the cpu field in the cross file takes no effect
and machine_args in config/arm/meson.build actually works.
Use a common one in cross files to remove any confusion.

Reported-by: Honnappa Nagarahalli <honnappa.nagarahalli@arm.com>
Signed-off-by: Joyce Kong <joyce.kong@arm.com>
Reviewed-by: Ruifeng Wang <ruifeng.wang@arm.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>