net/iavf: support Tx LLDP on scalar and AVX512
This patch adds a testpmd command "set tx lldp on" which will register
an mbuf dynflag at the application level, currently only supported on.

The IAVF will fills the SWTCH_UPLINK bit in the Tx context descriptor
based on the mbuf dynflag to send the LLDP packets.

For avx512, need to close the Tx port first, then "set tx lldp on",
and reopen the port to select correct path.

Signed-off-by: Zhichao Zeng <zhichaox.zeng@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>