net/bnxt: add support for backing store v2
Add backing store v2 changes.
The firmware supports the new backing store scheme for P7
and newer devices.

To support this, the driver queries the different types of chip
contexts the firmware supports and allocates the appropriate size
of memory for the firmware and hardware to use.
The code then goes ahead and frees up the memory during cleanup.

Older P5 device family continues to support the version 1 of
backing store. While the P4 device family does not need any
backing store memory.

Signed-off-by: Ajit Khaparde <ajit.khaparde@broadcom.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>