app/test: prevent exiting after skipped test
When processing a list of tests to run, the loop termination condition
was a test returning a value != 0. This means that if one of a series of
tests was skipped, i.e. returned TEST_SKIPPED, the whole execution run
was stopped. Since a test being skipped is not an error condition, we
put in an explicit check for that to keep executing any remaining tests.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>