net/tap: add deprecation notice for RTE_FLOW_ACTION_RSS
The BPF infrastructure in kernel changed which broke the method
being used to load the RSS part of rte_flow in the TAP device.
Since this is a little used feature, and fixing the problem would
be moderately complex and create new dependency on libbpf,
mark the feature as deprecated and plan for removal.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Acked-by: Ferruh Yigit <ferruh.yigit@amd.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>