arg_parser: add new coremask parsing API
Add new coremask parsing API. This API behaves similarly
to the corelist parsing API, taking a coremask string, a cores
array and a cores_len int. Parsing the coremask string, filling
its values into the cores array up to cores_len.

The API also returns a 'count' which corresponds to the total number
of cores in the coremask string.

Signed-off-by: Euan Bourke <euan.bourke@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>