dts: add required methods to testpmd_shell
Added a method within the testpmd interactive shell that polls the
status of ports and verifies that the link status on a given port is
"up." Polling will continue until either the link comes up, or the
timeout is reached. Also added methods for starting and stopping packet
forwarding in testpmd and a method for setting the forwarding mode on
testpmd. The method for starting packet forwarding will also attempt to
verify that forwarding did indeed start by default.

Signed-off-by: Jeremy Spewock <jspewock@iol.unh.edu>
Signed-off-by: 0-day Robot <robot@bytheb.org>