doc: updated incorrect value for IP frag max fragments
Docs for IP Fragment said RTE_LIBRTE_IP_FRAG_MAX_FRAGS was 4 by default,
however this was changed to 8.

Documentation has been updated to account for this, including a
snippet of the code where RTE_LIBRTE_IP_FRAG_MAX_FRAGS is defined to
ensure the documentation stays up to date.

Signed-off-by: Euan Bourke <euan.bourke@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>